% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnivGrChEdge.r
\name{Edge.dens}
\alias{Edge.dens}
\title{Univariate Edgeworth density}
\usage{
Edge.dens(w, n = 1, cum)
}
\arguments{
\item{w}{a vector of data,}

\item{n}{number of terms in the  mean}

\item{cum}{c(muX,varX,skewX,kurtX) the cumulants of the elements of the sum}
}
\value{
Edgeworth approximation evaluated at sample points w. w can be considered
a mean of \eqn{n_0} independent copies of a rv X.
}
\description{
Fits the Edgeworth approximation
}
\details{
Computes the Edgeworth density at the points  \eqn{w = \bar{x}_{n_0}= (X_1+ \dots + X_{n_0}}) where
\eqn{X_1, \dots, X_{n_0}} are i.i.d. rv with mean \eqn{\mu_X} and
variance \eqn{\sigma_X^2}
\deqn{
f_{{W}^{\left( n_0\right) }}\left( {w}\right) = \left( 1+\sum_{k=1}^{2}\frac{n_0^{-k/2}}{k!}g_{{Y},k}\left(
{z}\right) \right) \varphi \left( {w}|\mu_X,{\sigma }^2_{%
{X}}\right)
 }
where
\deqn{
g_{{Y},1}\left( {z}\right) =\frac{{\kappa }_{%
{Y},3}{H}_{3}\left( {{z}}\right) }{6}, \qquad
g_{{Y},2}\left( {z}\right) =\frac{{\kappa }_{%
 {Y},4}{H}_{4}\left( {{z}}\right) }{12}+\frac{{\kappa }_{{Y},3}^{ 2
 }{H}_{6}\left( {{z}}\right) }{36}.
}
Here \eqn{Y = (X-\mu_X)/\sigma_X} and \eqn{Z=\sqrt{n_0}(w-\mu_X)/\sigma_X}
}
\examples{
## GH distribution example
param <- c(mu=0, delta= 2, alpha= 1.8, beta= -0.91, lambda=2)
## generate GH random data
set.seed(123)
xGH <- GeneralizedHyperbolic::rghyp(1000,param=param)
xGH <- sort(xGH)

## sample cumulants values
Scum <- MultiStatM::SampleMomCum(xGH,r=4,centering=FALSE,scaling=FALSE)$estCum.r
## Mean, variance, sekwness and excess kurtosis
ScumGH <- c(Scum[1],Scum[2],Scum[3]/Scum[2]^(3/2),Scum[4]/Scum[2]^(4/2))
## Fit Edgeworth density n=1
EdgeFit <- Edge.dens(xGH, n=1, ScumGH)
## plot results
hist(xGH, freq=FALSE, breaks=20, main = "xGH data")
lines(xGH,EdgeFit, col="red",lwd=2)


}
