% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnivGrChEdge.r
\name{Edge.dens.squ}
\alias{Edge.dens.squ}
\title{Squared Edgeworth density}
\usage{
Edge.dens.squ(w, cum)
}
\arguments{
\item{w}{data vector}

\item{cum}{c(muX,varX,skewX,kurtX)}
}
\value{
density evaluate at w
}
\description{
Fits the Edgeworth approximation with squared polynomial part.
Only the case \eqn{n_0=1} is considered. Parameters are estimated
by \code{\link[=Edge.squ.MLE]{Edge.squ.MLE()}}
}
