% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnivGrChEdge.r
\name{Edge.Int}
\alias{Edge.Int}
\title{Integrated  Edgeworth density}
\usage{
Edge.Int(x, cum)
}
\arguments{
\item{x}{vector of upper limits of integration}

\item{cum}{vector (mean, variance, skewness, ex-kurtosis)}
}
\value{
the vector of estimated probabilities
}
\description{
Provide the integral for the Edgeworth density truncated at k=4
}
\details{
Provide the integral for the Edgeworth \code{\link[=Edge.dens]{Edge.dens()}} density truncated
at k=4 over the region (-infinity,x)
}
\examples{
## GH distribution example
param <- c(mu=0, delta= 2, alpha= 1.8, beta= -0.91, lambda=2)
## generate GH random data
set.seed(123)
xGH <- GeneralizedHyperbolic::rghyp(1000,param=param)
xGH <- sort(xGH)

## sample cumulants values
Scum <- MultiStatM::SampleMomCum(xGH,r=4,centering=FALSE,scaling=FALSE)$estCum.r
## Mean, variance, sekwness and excess kurtosis
ScumGH <- c(Scum[1],Scum[2],Scum[3]/Scum[2]^(3/2),Scum[4]/Scum[2]^(4/2))
## estimate probabilities on (-infinity,q)
q <- c(-1, -2, 0, 1)
Edge.Int(q, ScumGH)

}
