% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnivGrChEdge.r
\name{GrCh.dens.squ}
\alias{GrCh.dens.squ}
\title{Squared Gram-Charlier density of order K}
\usage{
GrCh.dens.squ(x, param)
}
\arguments{
\item{x}{vector of evaluation points}

\item{param}{c(mu, s2, alpha3, alpha4, ..., alphaK)}
}
\value{
the values of squared GC evaluated at x
}
\description{
Fits the squared polynomial Gram-Charlier density of order K
}
\details{
\deqn{
f_{{X}}\left( {x}\right) =\left(
1+\sum_{k=3}^{\infty }\frac{1}{k!}{B}_{k}\left( 0,0,%
{\kappa }_{{Y},3},\ldots {\kappa }_{%
{Y},k}\right) {H}_{k}\left( {y}|{I}%
 \right) \right)^2 \varphi \left( {x}|{\mu}_X,{\sigma }_X^2\right).
}

The parameters are estimated by \code{\link[=GrCh.squ.MLE]{GrCh.squ.MLE()}} or by \code{\link[=GrCh.minE]{GrCh.minE()}}.
}
\examples{
## GH distribution example
param <- c(mu=0, delta= 2, alpha= 1.8, beta= -0.91, lambda=2)
## generate GH random data
set.seed(123)
xGH <- GeneralizedHyperbolic::rghyp(1000,param=param)
xGH <- sort(xGH)
## MLE estimation
MLE <- GrCh.squ.MLE(data=xGH,start=c(0,1,0,3), method="BFGSR")
## fit the squared Gram Charlier
GCSqu <- GrCh.dens.squ(xGH, MLE$estimate)
## plot results
hist(xGH, freq=FALSE, breaks=20, main = "xGH data")
lines(xGH, GCSqu,  col = "#1E90FF", lwd = 2, lty = 2)

}
