% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnivGrChEdge.r
\name{GrCh.Int.squ}
\alias{GrCh.Int.squ}
\title{Integrated Squared Gram Charlier density}
\usage{
GrCh.Int.squ(x, param)
}
\arguments{
\item{x}{vector of values}

\item{param}{vector c(mu, sigma2, a3, ... ak)}
}
\description{
Provides the integral for any order k of the suared polynomial density
}
\details{
\deqn{
f_{{X}}\left( {x}\right) =\left(
1+\sum_{k=3}^{\infty }\frac{1}{k!}{B}_{k}\left( 0,0,%
{\kappa }_{{Y},3},\ldots {\kappa }_{%
{Y},k}\right) {H}_{k}\left( {y}|{I}%
 \right) \right)^2 \varphi \left( {x}|{\mu}_X,{\sigma }_X^2\right).
}

The parameters are estimated by \code{\link[=GrCh.squ.MLE]{GrCh.squ.MLE()}} or \code{\link[=GrCh.minE]{GrCh.minE()}}
}
