% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnivGrChEdge.r
\name{GrCh.squ.MLE}
\alias{GrCh.squ.MLE}
\title{MLE estimation of squared Gram Charlier density}
\usage{
GrCh.squ.MLE(data, start, method = "BFGSR", ...)
}
\arguments{
\item{data}{vector of data}

\item{start}{c(mu, sig2, alpha3, alpha4)}

\item{method}{optimization methods available in maxLik}

\item{...}{Additional arguments passed to maxLik()}
}
\value{
an object from maxLik
}
\description{
it uses maxLik to produce MLE estimates for \code{\link[=GrCh.dens.squ]{GrCh.dens.squ()}}
}
\examples{
## GH distribution example
param <- c(mu=0, delta= 2, alpha= 1.8, beta= -0.91, lambda=2)
## generate GH random data
set.seed(123)
xGH <- GeneralizedHyperbolic::rghyp(1000,param=param)
xGH <- sort(xGH)
## MLE estimation
MLE <- GrCh.squ.MLE(data=xGH,start=c(0,1,0,3), method="BFGSR")
## fit the squared Gram Charlier
GCSqu <- GrCh.dens.squ(xGH, MLE$estimate)
## plot results
hist(xGH, freq=FALSE, breaks=20, main = "xGH data")
lines(xGH, GCSqu,  col = "#1E90FF", lwd = 2, lty = 2)


}
